Fix an issue bundling our copy of the Backbone types.
We were using the dts rollup plugin when attempting to bundle our fork
of the Backbone types, but this plugin ends up rewriting those. This is
a problem for the particular setup in the Backbone types file. What we
really wanted to do was just copy it as-is.

This change does exactly that. We now use `rollup-plugin-copy` to just
simply copy anything in `src/@types` into `lib`, preserving it as-is.

This fixes issues with using Spina with TypeScript when using the built
package.

Testing Done:
Built a package and consumed it successfully without typing issues.

Reviewed at https://reviews.reviewboard.org/r/12794/