Mark Backbone and underscore as externals, not globals.
This updates the Rollup configuration for Spina to take in `Backbone`
and `underscore` as externals. This will pass in `Backbone` and `_` in
the UMD wrapper for the module, taking them the root object (such as
`window`) or from the calling module.

Defining these as externals is important for module resolution, and to
avoid the assumption of what provides these modules.

Testing Done:
Verified that the resulting UMD wrapper takes in `Backbone` and `_` as
parameters and that they're populated automatically.

Also verified that I could successfully import these into a module
bundling this with Backbone, and that resolution worked correctly.

Reviewed at https://reviews.reviewboard.org/r/12796/