Fix a casing issue with the referenced Backbone imports.
When originally building Spina, I started off with some Backbone on
TypeScript tutorial that had imports from `'Backbone'`. This worked fine
and I went with it. However, this was the wrong casing for the module,
which should have been `'backbone'`.

The reason this wasn't caught is that I'm developing on macOS, which has
case-insensitive filenames. Likely, so was the author of the tutorial.
Upon using this on a Linux system, imports failed.

This change updates all imports and the TypeScript configuration to use
the correct casing for the module.

Testing Done:
Successfully built this with lowercase `backbone`.

Verified that the new builds corrected build issues on Linux.

Reviewed at https://reviews.reviewboard.org/r/12817/