Add base support for running Jasmine unit tests.
This adds the beginnings of unit test support for Spina. We now have a
`npm run test` command that will invoke Jasmine on our test suite,
verifying that we don't regress and that everything behaves as expected.

For DOM-related tests (needed for `BaseView`), we make use of `jsdom`,
which simulates a DOM. This works along with jQuery. An in-memory
environment is set up before tests are run. No browser environment is
required for this.

Currently, there are no tests, but those will be coming along with some
in-progress work to fix and tighten up base object support.

Testing Done:
Tested along with in-progress unit tests, which pass.

Reviewed at https://reviews.reviewboard.org/r/12910/