Fix inheriting prototypeAttrs and automergeAttrs.
When support for `prototypeAttrs` and `automergeAttrs` were added, the
intent was for these to be inherited in any subclasses, even if
overridden. However, some logic issues prevented this from working
correctly.

The values from either the parent or subclass's lists were only being
populated if we found the value explicitly set in the subclass. And then
if a subclass set a list, it would likely override the parent's.

We now track any unique values we know we want copied into the subclass,
regardless of whether we copied the corresponding attribute values,
giving us a comprehensive final list.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13217/