Make onRender() and onInitialRender() protected, not public.
When first designing Spina, we made `onRender()` and `onInitialRender()`
public methods, when they really should have been protected methods,
discouraging any use outside of subclasses.

This changes them to protected. As this is an interface change, we'll be
making this release 3.0 instead of 2.1. That's a better fit for some of
the other changes introduced in this release anyway. A future change
will formalize the version bump.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/13550/