Fix a missing Backbone import, leading to broken types.
Spina 3.0 introduced a `Result<T> = Backbone._Result<T>` alias in
`objects.ts`, but it failed to import `Backbone`. This led to an
assumption that `Backbone` was a global type that should not be
interfered with, and thus renamed the `Backbone` type in the resulting
`d.ts` file tto `Backbone$1`.

This broke typing in some ways. `Result<T>` no longer had a type (since
it didn't know what `Backbone._Result<T>` was in this context), and the
difference between `Backbone` and `Backbone$1` led to issues resolving
types of some attributes in subclasses of our base classes.

This change adds the missing import, fixing up the types.

Testing Done:
Unit tests pass.

Verified this fixed typing issues in Ink.

Reviewed at https://reviews.reviewboard.org/r/13586/